requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
charset-normalizer==2.1.1
discord.py==2.1.0
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
imageio-ffmpeg==0.4.7
jmespath==1.0.1
multidict==6.0.3
numpy==1.23.5
pycord==0.1.1
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
yarl==1.8.2
